Update README.md
Select research articles and read or extract the text of the introduction portion. 
Preprocessing the introduction portion, remove the reference sentences. 
Apply the pre-train model to classify the whole generated sentence into two groups (positive and negative)
The positive sentence, remove and ignore. Focus on negative sentences.
I applied the extractive summarization techniques for negative sentences to generate a summary of each article.
Generates triples of the summary sentences.